Remove articles, json-request; Rename /chat -> /ticket, /chatApp -> /chat
