--- public/modules/articles/articles.client.module.js
@@ -1,4 +1,0 @@
-'use strict';
-
-// Use Application configuration module to register a new module
-ApplicationConfiguration.registerModule('articles');
