--- public/modules/articles/services/articles.client.service.js
@@ -1,14 +1,0 @@
-'use strict';
-
-//Articles service used for communicating with the articles REST endpoints
-angular.module('articles').factory('Articles', ['$resource',
-		function($resource) {
-			return $resource('articles/:articleId', {
-				articleId: '@_id'
-			}, {
-				update: {
-					method: 'PUT'
-				}
-			});
-		}
-]);
