--- public/modules/chat-app/chat-app.client.module.js
@@ -1,10 +1,0 @@
-'use strict';
-
-ApplicationConfiguration.registerModule('chatApp');
-
-angular.module('chatApp')
-.value('messageFormatter', function(date, nick, message) {
-	return date.toLocaleTimeString() + ' - ' + 
-		nick + ' - ' + 
-		message + '\n';
-});
