--- public/modules/chat-app/views/chat-app.client.view.html
@@ -1,32 +1,0 @@
-<div class='container'>
-  <div ng-controller='SocketCtrl'>
-    You are known as: <span ng-bind="nickName"></span>.
-    <br/>
-    Change your nickname with <code>/nick [yourNewNick]</code>
-    <div class='row'>
-      <div class='col-md-9'>
-        <textarea style='width: 100%; height: 200px'
-             ng-disable='true' ng-model='messageLog'></textarea>
-      </div>
-    </div>
-    <div class='row'>
-      &nbsp;
-    </div>
-    <div class='row'>
-      <div class='col-md-8'>
-        <form role='form' novalidate name='form'>
-          <div class='form-group'>
-            <input type='text' style='width: 80%'
-            required
-            ng-max='50'
-            ng-model='message'
-            focus>
-            <button class='btn btn-sm btn-primary'
-              ng-enable='form.$valid'
-              ng-click="sendMessage(message)">Send!</button>
-          </div>
-        </form>
-      </div>
-    </div>
-  </div>
-</div>
