+++ public/modules/chat/chat.client.module.js
@@ -1,3 +1,10 @@
 'use strict';
 
 ApplicationConfiguration.registerModule('chat');
+
+angular.module('chat')
+.value('messageFormatter', function(date, nick, message) {
+	return date.toLocaleTimeString() + ' - ' + 
+		nick + ' - ' + 
+		message + '\n';
+});
