+++ public/modules/core/views/header.client.view.html
@@ -15,7 +15,7 @@
             <li><a ui-sref="signin" ng-show="!loggedIn">Log In</a></li>
 						<li><a ui-sref="signup" ng-show="!loggedIn">Sign Up</a></li>
 						<li><a href="/#!/servers" ng-show="loggedIn">Current Tickets</a></li>
-						<li><a href="/#!/chat" ng-show="loggedIn">New Ticket</a></li>
+						<li><a href="/#!/ticket" ng-show="loggedIn">New Ticket</a></li>
 						<li><a href="/auth/signout" ng-show="loggedIn">Sign Out</a></li>
 						<li data-ng-bind="authentication.user.email"></li>
