--- public/modules/json-request/controllers/ticket.client.controller.js
@@ -1,8 +1,0 @@
-'use strict';
-
-angular.module('json-request').controller('TicketController', ['$scope',
-	function($scope) {
-		// Controller Logic
-		// ...
-	}
-]);
