--- public/modules/json-request/json-request.client.module.js
@@ -1,4 +1,0 @@
-'use strict';
-
-// Use application configuration module to register a new module
-ApplicationConfiguration.registerModule('json-request');
