--- public/modules/json-request/services/jsonrequest.client.service.js
@@ -1,15 +1,0 @@
-'use strict';
-
-angular.module('json-request').factory('Jsonrequest', [
-	function() {
-		// Jsonrequest service logic
-		// ...
-
-		// Public API
-		return {
-			someMethod: function() {
-				return true;
-			}
-		};
-	}
-]);
