--- public/modules/json-request/views/json-request.client.view.html
@@ -1,4 +1,0 @@
-<section data-ng-controller="JsonRequestController">
-				{{status}}
-				{{data}}	
-</section>
