--- public/modules/json-request/views/ticket.client.view.html
@@ -1,3 +1,0 @@
-<section data-ng-controller="TicketController">
-	Fuck
-</section>
