+++ public/modules/ticket/css/overrides/override.css
