+++ public/modules/ticket/ticket.client.module.js
@@ -1,0 +1,3 @@
+'use strict';
+
+ApplicationConfiguration.registerModule('ticket');
